refactor(ffmpeg): reorganize utility functions into utils subdirectory
- Move buildCodecOptions.ts from ffmpeg/ to ffmpeg/utils/
- Move buildFilter.ts from ffmpeg/ to ffmpeg/utils/
- Create utils/index.ts barrel export for buildCodecOptions and buildFilter
- Update getTrackSettings.ts to import from utils instead of relative paths
