+++ src/main/lib/ffmpeg/utils/buildCodecOptions.ts
