+++ src/main/lib/ffmpeg/utils/buildFilter.ts
@@ -1,0 +1,28 @@
+/*
+ * sound-balance-electron
+ * Copyright (C) 2026 Pavel Alloyarov
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+export const buildFilter = (
+	filterName: string,
+	options: Record<string, string | number | boolean>,
+) => {
+	const optionsArray = Object.entries(options);
+	if (optionsArray.length > 0) {
+		const filterArgs = optionsArray.map(([k, v]) => `${k}=${v}`).join(":");
+		return ["-af", `${filterName}=${filterArgs}`];
+	}
+	return ["-af", `${filterName}`];
+};
